+++ crispy_forms/templatetags/crispy_forms_tags.py
@@ -1,20 +1,19 @@
-# -*- coding: utf-8 -*-
 from django import template
 from django.conf import settings
 from django.forms.formsets import BaseFormSet
 from django.template.loader import get_template
 
-from crispy_forms.compatibility import lru_cache, string_types
+from crispy_forms.compatibility import lru_cache
 from crispy_forms.helper import FormHelper
 from crispy_forms.utils import TEMPLATE_PACK, get_template_pack
 
 register = template.Library()
 
 # We import the filters, so they are available when doing load crispy_forms_tags
-from crispy_forms.templatetags.crispy_forms_filters import *  # isort:skip
+from crispy_forms.templatetags.crispy_forms_filters import *  # NOQA: F403,F401 isort:skip
 
 
-class ForLoopSimulator(object):
+class ForLoopSimulator:
     """
     Simulates a forloop tag, precisely::
 
@@ -253,7 +252,7 @@ def do_uni_form(parser, token):
     # {% crispy form 'bootstrap' %}
     if (
         helper is not None and
-        isinstance(helper, string_types) and
+        isinstance(helper, str) and
         ("'" in helper or '"' in helper)
     ):
         template_pack = helper