+++ crispy_forms/tests/test_utils.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals
-
 import pytest
 
 import django
@@ -24,8 +21,7 @@ def test_render_field_with_none_field():
     rendered = render_field(field=None, form=None, form_style=None, context=None)
     assert rendered == ''
 
-@pytest.mark.skipif(django.VERSION < (1, 9),
-                    reason="Custom BoundField behavior is was introduced in 1.9.")
+
 def test_custom_bound_field():
     from django.forms.boundfield import BoundField
 
@@ -44,7 +40,7 @@ def test_custom_bound_field():
         f = MyCharField()
 
         def __init__(self, *args, **kwargs):
-            super(MyForm, self).__init__(*args, **kwargs)
+            super().__init__(*args, **kwargs)
             self.helper = FormHelper()
             self.helper.layout = Layout('f')