+++ crispy_forms/utils.py
@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 import logging
 import sys
 
@@ -10,7 +8,7 @@ from django.template.loader import get_template
 from django.utils.functional import SimpleLazyObject
 
 from .base import KeepContext
-from .compatibility import PY2, text_type, lru_cache
+from .compatibility import lru_cache
 
 
 def get_template_pack():
@@ -60,20 +58,6 @@ def render_field(
             return field.render(
                 form, form_style, context, template_pack=template_pack,
             )
-        else:
-            # In Python 2 form field names cannot contain unicode characters without ASCII mapping
-            if PY2:
-                # This allows fields to be unicode strings, always they don't use non ASCII
-                try:
-                    if isinstance(field, text_type):
-                        field = field.encode('ascii').decode()
-                    # If `field` is not unicode then we turn it into a unicode string, otherwise doing
-                    # str(field) would give no error and the field would not be resolved, causing confusion
-                    else:
-                        field = text_type(field)
-
-                except (UnicodeEncodeError, UnicodeDecodeError):
-                    raise Exception("Field '%s' is using forbidden unicode characters" % field)
 
         try:
             # Injecting HTML attributes into field's widget, Django handles rendering these
@@ -109,7 +93,7 @@ def render_field(
                 logging.warning("Could not resolve form field '%s'." % field, exc_info=sys.exc_info())
 
         if hasattr(form, 'rendered_fields'):
-            if not field in form.rendered_fields:
+            if field not in form.rendered_fields:
                 form.rendered_fields.add(field)
             else:
                 if not FAIL_SILENTLY: